+++ src/Forcing/Forcing.jl
@@ -5,7 +5,9 @@ Defines methods for injecting energy onto a system of vortices.
 """
 module Forcing
 
+using ..Filaments: AbstractFilament, UnitTangent
 using ..SyntheticFields: SyntheticFields, SyntheticVectorField
+using LinearAlgebra: ×
 
 export AbstractForcing, NormalFluidForcing
 
@@ -15,6 +17,13 @@ export AbstractForcing, NormalFluidForcing
 Abstract type representing a forcing method.
 """
 abstract type AbstractForcing end
+
+"""
+    Forcing.apply!(forcing::AbstractForcing, vs::AbstractVector, f::AbstractFilament)
+
+Apply forcing to a single filament `f`, modifying vortex velocities `vs`.
+"""
+function apply! end
 
 @doc raw"""
     NormalFluidForcing <: AbstractForcing
@@ -91,10 +100,34 @@ end
 
 function Base.show(io::IO, f::NormalFluidForcing{T, N}) where {T, N}
     (; vn, α, α′,) = f
+    indent = get(io, :indent, 0)
+    nspaces = max(indent, 1)
+    spaces = " "^nspaces
     print(io, "NormalFluidForcing{$T, $N} with:")
-    print(io, "\n - Magnus force coefficient: α = ", α)
-    print(io, "\n - Drag force coefficient: α′ = ", α′)
-    print(io, "\n - Normal velocity field: ", vn)
+    print(io, "\n$(spaces)├─ Magnus force coefficient: α = ", α)
+    print(io, "\n$(spaces)├─ Drag force coefficient: α′ = ", α′)
+    print(io, "\n$(spaces)└─ Normal velocity field: ", vn)
+end
+
+function apply!(forcing::NormalFluidForcing, vs::AbstractVector, f::AbstractFilament)
+    length(vs) == length(f) || throw(DimensionMismatch("length of filament and velocity vector don't match"))
+    (; vn, α, α′,) = forcing
+    V = eltype(vs)  # usually Vec3{T} = SVector{3, T}
+    # TODO: parallelise?
+    for i ∈ eachindex(vs, f)
+        vs_i = vs[i]  # vortex velocity before mutual friction forcing
+        s⃗ = f[i]                  # vortex point location
+        s⃗′ = f[i, UnitTangent()]  # local tangent
+        vn_i = V(vn(s⃗))        # evaluate normal fluid velocity at s⃗
+        v_ns = vn_i - vs_i     # slip velocity
+        v_ns_perp = s⃗′ × v_ns  # slip velocity orthogonal to local tangent
+        vf = α * v_ns_perp     # velocity due to Magnus force
+        if !iszero(α′)
+            vf = vf - α′ * s⃗′ × v_ns_perp  # velocity due to drag force (it's quite common to set α′ = 0)
+        end
+        vs[i] = vs[i] + vf
+    end
+    vs
 end
 
 end