+++ CHANGELOG.md
@@ -7,8 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `VortexPasta.Forcing` module allowing to specify "forcing" methods.
+  Currently, this includes the `NormalFluidForcing` type for modifying the
+  vortex velocities via mutual friction with a (synthetic) normal fluid.
+  Moreover, the `init` function (in `Timestepping`) now accepts a `forcing`
+  argument which can be used in simulations.
+
 - Add `VortexPasta.SyntheticFields` module allowing to create and evaluate
-  synthetic vector fields (which will be used in the future to represent a normal fluid).
+  synthetic vector fields (such as a normal fluid velocity field).
+
 
 ## [0.24.8] - 2024-11-14